server/requirements.txt
# Серверные зависимости для VPS 109.94.211.29.
# Устанавливаются в /opt/planner/venv на сервере.
# Клиент (CustomTkinter и т.д.) живёт в корневом requirements.txt — не здесь.

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
aiogram>=3.15.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
alembic>=1.13.0
slowapi>=0.1.9
httpx>=0.27.0
tzdata>=2024.1  # zoneinfo timezone data для Windows и других не-Linux платформ
